Consolidate S3 Integrations
