+++ src/app.css
@@ -281,6 +281,11 @@ hr {
   text-decoration: underline;
 }
 
+.s3-modified {
+  margin-left: 2px;
+  color: grey;
+}
+
 @media screen and (max-width: 1054px) {
   .home .card:nth-of-type(2){
     margin-top: 2em;
