+++ src/data-publication/reports/DynamicDataset.css
@@ -35,8 +35,3 @@
 .DynamicDataset .item:last-child {
   margin-right: 0;
 }
-
-.DynamicDataset #datasetList .modified {
-  margin-left: 2px;
-  color: grey;
-}
