+++ src/data-publication/reports/snapshot/Snapshot.css
@@ -39,8 +39,3 @@
 .Snapshot .item li > ul {
   margin-top: 0.5em;
 }
-
-.Snapshot #datasetList .modified {
-  margin-left: 2px;
-  color: grey;
-}
