+++ src/documentation/index.css
@@ -49,7 +49,3 @@
 .App.Documentation li {
   margin-bottom: 0.5em;
 }
-
-.Documentation .s3-modified {
-  color: grey;
-}
